% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{calc_corr}
\alias{calc_corr}
\title{Correlation between a vector and time}
\usage{
calc_corr(x, t)
}
\arguments{
\item{x}{Numeric vector.}

\item{t}{Time vector.}
}
\value{
Numeric correlation.
}
\description{
Correlation between a vector and time
}
