% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasySILAC-package.R
\docType{package}
\name{EasySILAC-package}
\alias{EasySILAC}
\alias{EasySILAC-package}
\title{EasySILAC: SILAC protein turnover analysis with R and Rust fitting engines}
\description{
Fitting (steady-state and non-steady-state), labeling-efficiency (Hill)
correction, QC and visualization for pulsed SILAC proteomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yuanlizhanshi/EasySILAC}
  \item Report bugs at \url{https://github.com/yuanlizhanshi/EasySILAC/issues}
}

}
\author{
\strong{Maintainer}: Yunhui Kong \email{kongyunhui1@gamil.com}

Authors:
\itemize{
  \item Yunhui Kong \email{kongyunhui1@gamil.com}
}

}
\keyword{internal}
