% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{extract_ntr_from_se}
\alias{extract_ntr_from_se}
\title{Extract the new-to-total ratio of one peptide}
\usage{
extract_ntr_from_se(se, peptide)
}
\arguments{
\item{se}{SummarizedExperiment.}

\item{peptide}{Peptide row name.}
}
\value{
A tibble with time and ratio.
}
\description{
Extract the new-to-total ratio of one peptide
}
