% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{get_weighted_turnover}
\alias{get_weighted_turnover}
\title{Turnover weighted by a quantitative assay}
\usage{
get_weighted_turnover(protein_fit_se, assay = "ratio")
}
\arguments{
\item{protein_fit_se}{SummarizedExperiment with fit results in `rowData`.}

\item{assay}{Assay used for weighting (default "ratio").}
}
\value{
A tibble of weighted turnover statistics.
}
\description{
Turnover weighted by a quantitative assay
}
