% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg}
\alias{msg}
\title{Print a timestamped message}
\usage{
msg(..., verbose = TRUE)
}
\arguments{
\item{...}{Passed to `cat()`.}

\item{verbose}{Logical; if FALSE the message is suppressed.}
}
\value{
Invisibly NULL.
}
\description{
Print a timestamped message
}
