% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_density}
\alias{plot_density}
\title{Density plot of a numeric column}
\usage{
plot_density(data, column = NULL, title = NULL, add_2_fold = FALSE)
}
\arguments{
\item{data}{Data frame.}

\item{column}{Column name to plot.}

\item{title}{Plot title.}

\item{add_2_fold}{Add dashed lines at two-fold changes.}
}
\value{
A ggplot object.
}
\description{
Density plot of a numeric column
}
