% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pep}
\alias{plot_pep}
\title{Plot one peptide's ratio time course}
\usage{
plot_pep(se, pep)
}
\arguments{
\item{se}{SummarizedExperiment.}

\item{pep}{Peptide row name.}
}
\value{
A ggplot object.
}
\description{
Plot one peptide's ratio time course
}
