% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{geom_mean}
\alias{df2mtx}
\title{EasyProtein re-exports}
\usage{
geom_mean(x, na.rm = TRUE)

df2mtx(df)
}
\description{
`geom_mean` and `df2mtx`, re-exported from \pkg{EasyProtein}.
}
\keyword{internal}
