% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_plot}
\alias{remove_plot}
\title{Remove saved plot files}
\usage{
remove_plot(filename)
}
\arguments{
\item{filename}{Base file name (without extension).}
}
\value{
Invisibly NULL.
}
\description{
Remove PNG/PDF/TIFF files previously written by [saveplot()].
}
