% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{savepng}
\alias{savepng}
\title{Save a plot to PNG only}
\usage{
savepng(object, filenames = filename, width = NULL, height = NULL, dpi = 600)
}
\arguments{
\item{object}{Plot object.}

\item{filenames}{Base file name.}

\item{width}{Width in inches (required).}

\item{height}{Height in inches (required).}

\item{dpi}{Resolution (default 600).}
}
\value{
Invisibly the written path.
}
\description{
Save a plot to PNG only
}
